requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
black @ git+git://github.com/psf/black@820f38708fd41a1b992716b1f65c9b0656f589d0
click==7.1.2
dataclasses==0.7
fn==0.4.3
gevent==20.6.2
greenlet==0.4.16
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
pylint==2.5.3
PyYAML==5.3.1
regex==2020.7.14
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
wrapt==1.12.1
zope.event==4.4
zope.interface==5.1.0
